Final Modification
Last version. There are still some issues (incl. a couple of typos) but overall it should work.
Recurring bug: Action menu shown by clicking on the selected item sometimes shows actions for other (previously selected) items, too. Tried to fix this issue in this last update and did not come across it in short testing of the final build - can't guaratee it is gone, though.
Recommendation is still to quit and re-launch the game from the executable, rather than clicking restart on the final screen. Potential for odd errors otherwise.